kconfig: let 'shell' return enough output for PATH_MAX
The 'shell' built-in only returns the first 256 bytes of the command's
output. In some cases, 'shell' is used to return a path; for instance
the gcc-plugins test added in 1e860048c53e ("gcc-plugins: simplify GCC
plugin-dev capability test") invokes '$(CC) -print-file-name=plugin'
which may be truncated in the case of a deeply-nested path.

Do the simple thing and bump up the buffer size to 4096; this is the
same value as PATH_MAX.

Signed-off-by: Brenda Streiff <brenda.streiff@ni.com>