net: dsa: mv88e6xxx: don't reject all timestamps if overwrite/discard bit set
If the overwritten or discard bits are set along with the timestamp, we
were treating the timestamp as invalid. This is overzealous; those bits
do tell us that timestamps have been lost, but the one that we did read
is a valid timestamp that is associated with a packet in the rx queue,
so we shouldn't drop it.

Signed-off-by: Brenda Streiff <brenda.streiff@emerson.com>