net: dsa: mv88e6xxx: embed rx timestamps into packets
Marvell's "Generation 3" PTP IP and later have support for embeddeding
received timestamps into incoming packets, which offers two advantages:

- no need to do SMI register accesses to retrieve the timestamps

- eliminates the possibility of having to overwrite/discard timestamps
  if they were not read in time.

On applicable hardware, we configure the switch to embed the timestamp
into the "reserved2" field (1588v3 calls this "messageTypeSpecific").
We still need to queue the packet for a worker because we need to take
the reg lock, but we can simplify down to only needing a single rx queue
instead of splitting based on PTP packet type.

Signed-off-by: Brenda Streiff <brenda.streiff@emerson.com>