net: dsa: mv88e6xxx: add 1588-related addresses to ATU as management frames
When packets ingress a trunked port, the DSA tag they receive will
typically indicate that they came from a trunk, and not the port. This is
undesirable in the case of protocols like 1588, where the timestamping
happens at a port, but when in a LAG-offloaded trunk, the DSA infrastructure
delivers the packet to the bond/team interface, which knows nothing about
timestamping.

This can be avoided with frames that the switch considers to be
"management"; therefore, define a set of "extra management DAs" to be
configured in the ATU. This set currently includes destination addresses
relevant for L2 and L4 1588.

This is not needed for 802.1AS, because 802.1AS packets use a destination
address of 01:80:c2:00:00:0e, which is already covered by the rsvd2cpu
functionality.

Signed-off-by: Brenda Streiff <brenda.streiff@emerson.com>